requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
pydantic==2.10.6
llama-cpp-python==0.3.4
chromadb==0.5.0
pypdf==4.2.0
python-dotenv==1.0.1
